Refine provider policy validation boundaries
